Tasks

Approved-by: Filip Konig <filip.konig39@gmail.com>

+++ src/app/components/table/DataTable.tsx
@@ -94,8 +94,8 @@ type Props = {
   setOrder: React.Dispatch<React.SetStateAction<OrderType>>;
   setGlobalFilter?: React.Dispatch<React.SetStateAction<string>>;
   onRowClick: (id: number) => void;
-  onCreate: () => void;
-  onDelete: (selectedRowIds: number[]) => void;
+  onCreate?: () => void;
+  onDelete?: (selectedRowIds: number[]) => void;
   maxCellLength?: number;
 };
 
@@ -141,38 +141,40 @@ const DataTable: React.FC<Props> = ({
     usePagination,
     useRowSelect,
     hooks => {
-      hooks.allColumns.push((_columns: any) => [
-        // Let's make a column for selection
-        {
-          id: 'selection',
-          // The header can use the table's getToggleAllRowsSelectedProps method
-          // to render a checkbox.  Pagination is a problem since this will select all
-          // rows even though not all rows are on the current page.  The solution should
-          // be server side pagination.  For one, the clients should not download all
-          // rows in most cases.  The client should only download data for the current page.
-          // In that case, getToggleAllRowsSelectedProps works fine.
-          // @ts-ignore
-          Header: ({ getToggleAllRowsSelectedProps }) => (
-            <div>
-              <IndeterminateCheckbox {...getToggleAllRowsSelectedProps()} />
-            </div>
-          ),
-          // The cell can use the individual row's getToggleRowSelectedProps method
-          // to the render a checkbox
-          // @ts-ignore
-          Cell: ({ row }: { row: any }) => (
-            <div>
-              <IndeterminateCheckbox
-                {...row.getToggleRowSelectedProps()}
-                onClick={(e: React.MouseEvent<HTMLInputElement>) =>
-                  e.stopPropagation()
-                }
-              />
-            </div>
-          ),
-        },
-        ..._columns,
-      ]);
+      if (onDelete) {
+        hooks.allColumns.push((_columns: any) => [
+          // Let's make a column for selection
+          {
+            id: 'selection',
+            // The header can use the table's getToggleAllRowsSelectedProps method
+            // to render a checkbox.  Pagination is a problem since this will select all
+            // rows even though not all rows are on the current page.  The solution should
+            // be server side pagination.  For one, the clients should not download all
+            // rows in most cases.  The client should only download data for the current page.
+            // In that case, getToggleAllRowsSelectedProps works fine.
+            // @ts-ignore
+            Header: ({ getToggleAllRowsSelectedProps }) => (
+              <div>
+                <IndeterminateCheckbox {...getToggleAllRowsSelectedProps()} />
+              </div>
+            ),
+            // The cell can use the individual row's getToggleRowSelectedProps method
+            // to the render a checkbox
+            // @ts-ignore
+            Cell: ({ row }: { row: any }) => (
+              <div>
+                <IndeterminateCheckbox
+                  {...row.getToggleRowSelectedProps()}
+                  onClick={(e: React.MouseEvent<HTMLInputElement>) =>
+                    e.stopPropagation()
+                  }
+                />
+              </div>
+            ),
+          },
+          ..._columns,
+        ]);
+      }
     },
   );
 
@@ -221,7 +223,9 @@ const DataTable: React.FC<Props> = ({
   }
 
   function handleDeleteMultiple() {
-    onDelete(selectedFlatRows.map((row: any) => row.original.id));
+    if (onDelete) {
+      onDelete(selectedFlatRows.map((row: any) => row.original.id));
+    }
   }
 
   const refetch = () => loadPage(0);
